feat(统一文章不同路径结构下的数据处理方式): 使得其他路径结构下的文章也能进行原生AI摘要处理
(cherry picked from commit f427696b37872db96de53a071818a3fa98c691cd)
